GeneratePDF class work
