+++ OcrExerciseGenerator-console/src/main/java/com/isa/jjdzr/App.java
@@ -1,11 +1,14 @@
 package com.isa.jjdzr;
 
+import com.isa.jjdzr.exercise.service.GeneratePDF;
+
 import static com.isa.jjdzr.exercise.service.GeneratePDF.generatePdf;
 
 public class App {
     public static void main(String[] args) {
-       JavaFanaticsTerminalLogo.javaFanaticsLogo();
-       Menu.menu();
+//       JavaFanaticsTerminalLogo.javaFanaticsLogo();
+//       Menu.menu();
+        GeneratePDF.generatePdf();
 
     }
 }
