+++ OcrExerciseGenerator-web/src/main/resources/templates/home.html
@@ -8,7 +8,7 @@
 <h1>Tutaj bedzie HomePage</h1>
 
 <h2>A tu Jacek testuje pdf</h2>
-<form method="POST" action="/generate-pdf">
+<form method="POST" action="http://localhost:8080/exercises/generate-pdf">
     <button type="submit">Generuj PDF</button></form>
 
 </body>
